requirements.txt
Flask
google-generativeai
requests
python-dotenv
markdown-it-py
bleach